@@ -1,22 +1,19 @@
 package com.jorgeacetozi.notepad.integrationTests.note.domain.service;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.Assert.assertThat;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import com.jorgeacetozi.notepad.note.domain.model.Note;
 import com.jorgeacetozi.notepad.note.domain.service.NoteService;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 @ActiveProfiles("test")
 public class NoteServiceTest {
@@ -26,12 +23,12 @@ public class NoteServiceTest {
 	
 	private Note note;
 	
-	@Before
+	@BeforeEach
 	public void setUp() {
 		note = new Note("Kubernetes", "Best container orchestration tool ever");
 	}
 	
-	@After
+	@AfterEach
 	public void destroy() {
 		noteService.delete(note);
 	}
