@@ -2,8 +2,8 @@ package com.jorgeacetozi.notepad.note.api;
 
 import java.util.List;
 
-import javax.validation.Valid;
 
+import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
